@@ -6,12 +6,10 @@ import { AppComponent } from './app.component';
 import { HomeModule } from './components/home/home.module';
 import { GeneralModule } from './components/general/general.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-// import { AnimateOnScrollModule } from 'ng2-animate-on-scroll';
 import { ServiceWorkerModule } from '@angular/service-worker';
 import { environment } from '../environments/environment';
-import { NgxGoogleAnalyticsModule } from 'ngx-google-analytics';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
-import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import {
   TranslateLoader,
   TranslateModule,
@@ -26,22 +24,19 @@ export function HttpLoaderFactory(http: HttpClient) {
 @NgModule({
   declarations: [
     AppComponent,
-    /* ArchiveComponent */
+  ],
+  bootstrap: [
+    AppComponent
   ],
   imports: [
     BrowserAnimationsModule,
-
     HomeModule,
     GeneralModule,
-
-    // AnimateOnScrollModule.forRoot(),
     BrowserModule,
     AppRoutingModule,
-    HttpClientModule,
     ServiceWorkerModule.register('ngsw-worker.js', {
-      enabled: environment.production,
+        enabled: environment.production,
     }),
-    // NgxGoogleAnalyticsModule.forRoot(environment.trackAnalyticID),
     TranslateModule.forRoot({
       loader: {
         provide: TranslateLoader,
@@ -49,9 +44,10 @@ export function HttpLoaderFactory(http: HttpClient) {
         deps: [HttpClient],
       },
     }),
-    NgbModule,
+    NgbModule
   ],
-  providers: [TranslateService],
-  bootstrap: [AppComponent],
+  providers: [
+    TranslateService, provideHttpClient(withInterceptorsFromDi())
+  ]
 })
 export class AppModule {}
